Split package into main.rs and lib.rs
